mejoras en vista pooles

+++ database/migrations/2025_03_11_150608_create_contratos_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('tecnologia')->nullable();
             $table->date('fecha_inicio');
             $table->date('fecha_fin')->nullable();
-            $table->decimal('precio', 8, 3);  // Monto de la factura o costo del plan
+            $table->bigInteger('precio');
             $table->timestamps();
         });
     }
